requirements.txt
pydantic-settings
python-dotenv
logfire